+++ src/chat-message/chat-item.tsx
@@ -80,19 +80,13 @@ export default class ChatItem extends Component<ChatMessageProps> {
   private getInternalMessage(): ChatMessagesData | null {
     const { message, role, content, status, id } = this.props;
 
-    // 优先使用直接传入的属性
-    if (role && content) {
-      const internalMessage: ChatMessagesData = {
-        id,
-        role,
-        content,
-        status,
-      } as ChatMessagesData;
-      return internalMessage;
-    }
-
-    // 兼容旧版本，使用 message 属性
-    return message || null;
+    // 优先使用直接传入的属性，这里不要先解构message再覆盖属性，发现会覆盖undefined
+    return {
+      id: id || message?.id,
+      role: role || message?.role,
+      content: content || message?.content,
+      status: status || message?.status,
+    } as ChatMessagesData;
   }
 
   receiveProps(
@@ -116,19 +110,13 @@ export default class ChatItem extends Component<ChatMessageProps> {
   ): ChatMessagesData | null {
     const { message, role, content, status, id = '' } = props;
 
-    // 优先使用直接传入的属性
-    if (role && content) {
-      const internalMessage: ChatMessagesData = {
-        id,
-        role,
-        content,
-        status,
-      } as ChatMessagesData;
-      return internalMessage;
-    }
-
-    // 兼容旧版本，使用 message 属性
-    return message || null;
+    // 优先使用直接传入的属性，这里不要先解构message再覆盖属性，发现会覆盖undefined
+    return {
+      id: id || message?.id,
+      role: role || message?.role,
+      content: content || message?.content,
+      status: status || message?.status,
+    } as ChatMessagesData;
   }
 
   ready(): void {
@@ -139,7 +127,7 @@ export default class ChatItem extends Component<ChatMessageProps> {
     const { name, datetime } = this.props;
     return (
       <slot name="header">
-        {!this.renderMessageStatus ? (
+        {!this.renderMessageWithStatus ? (
           <div class={`${className}__header`}>
             <span class={`${className}__name`}>
               <slot name="name">{this.renderNameContent(name)}</slot>
@@ -169,6 +157,9 @@ export default class ChatItem extends Component<ChatMessageProps> {
   }
 
   private renderAvatarContent(avatar: any) {
+    if (!avatar) {
+      return null;
+    }
     // 对于非字符串类型（包括React元素），都使用slot
     return isString(avatar) ? (
       <div className={`${className}__avatar-box`}>
@@ -195,13 +186,15 @@ export default class ChatItem extends Component<ChatMessageProps> {
     );
   };
 
-  get renderMessageStatus() {
+  get renderMessageWithStatus() {
     const internalMessage = this.getInternalMessage();
     if (!internalMessage) return null;
 
     const { status, content } = internalMessage;
     const { animation = 'skeleton' } = this.props;
-    if (status === 'pending' || (status === 'streaming' && content && content.length === 0)) {
+    const emptyContent = !content || content.length === 0;
+
+    if (status === 'pending' || (status === 'streaming' && emptyContent)) {
       return (
         <div class={`${className}-chat-loading`}>
           {convertToLightDomNode(
@@ -210,7 +203,9 @@ export default class ChatItem extends Component<ChatMessageProps> {
         </div>
       );
     }
-    // 这里不添加stop和error状态是避免影响已渲染内容
+    if (status === 'error' && (emptyContent || (content.length === 1 && content[0]?.type === 'text'))) {
+      return <div className={`${className}-chat-error`}>{content?.[0]?.data || '请求出错'}</div>;
+    }
     return null;
   }
 
@@ -346,11 +341,11 @@ export default class ChatItem extends Component<ChatMessageProps> {
       <div
         className={classname(baseClass, roleClass, variantClass, placementClass)}
         // 渲染出原生的header时要设置额外间隔，当用户slot自定义header时不管
-        data-has-header={!this.renderMessageStatus && (!!props.name || !!props.datetime)}
+        data-has-header={!this.renderMessageWithStatus && (!!props.name || !!props.datetime)}
       >
         {this.renderAvatar()}
-        {this.renderMessageStatus}
-        {!this.renderMessageStatus && (
+        {this.renderMessageWithStatus}
+        {!this.renderMessageWithStatus && (
           <div class={`${className}__main`}>
             {this.renderMessageHeader()}
             <slot name="content" className={`${className}__content__slot`}>