+++ src/upload/CustomFile.tsx
@@ -13,43 +13,14 @@ export interface CustomFileProps extends CommonDisplayFileProps {
   childrenNode?: VNode;
 }
 
-const CustomFile = (props: CustomFileProps) => {
+export function renderCustomFile(props: CustomFileProps) {
   const { classPrefix } = props;
-  // const drag = useDrag({ ...props.dragEvents, accept });
-  // const { dragActive } = drag;
-
-  // const dragEvents = draggable
-  //   ? {
-  //       onDrop: drag.handleDrop,
-  //       onDragEnter: drag.handleDragenter,
-  //       onDragOver: drag.handleDragover,
-  //       onDragLeave: drag.handleDragleave,
-  //     }
-  //   : {};
-
-  // const renderDragContent = () => {
-  //   const childrenContent = parseTNode(props.childrenNode, { dragActive, files: displayFiles });
-  //   const triggerContent = parseTNode(props.trigger, { dragActive, files: displayFiles });
-  //   return (
-  //     <div
-  //       className={`${classPrefix}-upload__dragger ${classPrefix}-upload__dragger-center`}
-  //       {...dragEvents}
-  //       onClick={props.triggerUpload}
-  //     >
-  //       <div className={`${classPrefix}-upload__trigger`}>
-  //         {parseTNode(props.dragContent, { dragActive, files: displayFiles }) || triggerContent || childrenContent}
-  //       </div>
-  //     </div>
-  //   );
-  // };
 
   return (
     <div className={`${classPrefix}-upload__trigger`} onClick={props.triggerUpload}>
       <slot></slot>
     </div>
   );
-};
+}
 
-CustomFile.displayName = 'CustomFile';
-
-export default CustomFile;
+export default renderCustomFile;